refactor(JWA): added JWA as a dependency to JOSE, removed the files with the original implemetation, removed export from this library
