+++ src/index.js
@@ -1,7 +1,6 @@
 /**
  * @module JSON Object Signing and Encryption (JOSE)
  */
-const JWA = require('./jose/JWA')
 const JWK = require('./jose/JWK')
 const JWKSet = require('./jose/JWKSet')
 const JWT = require('./jose/JWT')
@@ -18,7 +17,6 @@ const JWTSchema = require('./schemas/JWTSchema')
  * Export
  */
 module.exports = {
-  JWA,
   JWK,
   JWKSet,
   JWT,
